Move `scene.cpp` and `scene.h` to own subdirectory.
