+++ Scene/Scene.cpp
@@ -8,7 +8,7 @@ http://mozilla.org/MPL/2.0/.
 */
 
 #include "stdafx.h"
-#include "scene.h"
+#include "Scene/Scene.h"
 
 #include "simulation.h"
 #include "Globals.h"
